% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{check_codes}
\alias{check_codes}
\title{Find similar codes for a chosen period to avoid counting similar injury more than once.
Column \code{dup} with value 1 represents possibility of similar injury. The rows should
be carefully evaluated or deleted ie. delete all with \code{dup == 1}}
\usage{
check_codes(d, id = "lopenr", col = "hoveddiagnoser", days)
}
\arguments{
\item{d}{Dataset}

\item{id}{Unque id}

\item{col}{Column name for codes selection}

\item{days}{Condition for selection eg. 3 or 5 days}
}
\value{
A dataset with extra columname ie. \code{dup}. When \code{dup == 1} indicates the row is duplicated
with the specified period
}
\description{
Find similar codes for a chosen period to avoid counting similar injury more than once.
Column \code{dup} with value 1 represents possibility of similar injury. The rows should
be carefully evaluated or deleted ie. delete all with \code{dup == 1}
}
\examples{
\dontrun{
dj <- check_codes(dx, "lopenr", "hoveddiagnoser", 3)
}
}
\keyword{internal}
