% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{count_case}
\alias{count_case}
\title{Count cases}
\usage{
count_case(
  d,
  period = 0,
  id = "lopenr",
  date.col = "innDato",
  acute = FALSE,
  days = 0,
  diag.col = "hoveddiagnoser"
)
}
\arguments{
\item{d}{Dataset that contain ICD-10 diagnosis codes}

\item{period}{Representing 4-months period ie. first, second or third.
Default is 0 to include data for the whole period else use 1, 2 or 3.}

\item{id}{Columname representing unique id}

\item{date.col}{Columname for date for filtering}

\item{acute}{Default is \code{FALSE}. Use \code{TRUE} to include only acute patients
ie. Hastegrad = 1}

\item{days}{If diffence in days of registration should be considered ie. when
a person has more than one registered injuries of the same ICD-10 code.
This is to avoid counting similar injuries more than once.}

\item{diag.col}{Columname of codes for main diagnosis}
}
\value{
When the value for arg \code{days} is other than \code{0}, a new column called
\code{dup} is created ie. duplicated cases. Hence \code{dup == 1} indicates the row is
duplicated as defined within the specified days.
}
\description{
Count injury cases or episodes. Generally an episode shoulde be:
\itemize{
\item Main diagnosis is between S00 to T78
\item It's an acute injury ie. hastegrad is 1
\item Posibility to select days from previous to the following injury of similar
code for multiple injuries
}
}
\examples{
\dontrun{
dd <- count_case(dt1, period = 1:2, acute = TRUE)
dd <- count_case(dt1, acute = TRUE, days = 3)
}
}
