% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{find_case}
\alias{find_case}
\title{Find cases}
\usage{
find_case(
  d1,
  d2,
  id = "lopenr",
  skade = "skadeDato",
  rhf = "helseforetak_nr",
  filter = NULL,
  days = 3,
  verbose = FALSE,
  clean = FALSE
)
}
\arguments{
\item{d1}{Dataset for FMDS ie. external explanation for injury}

\item{d2}{Dataset for NPR ie. intery registration for injury}

\item{id}{Unique patient identity}

\item{skade}{Coloumn name for date of injury}

\item{rhf}{Coloumn name for health institutions}

\item{filter}{Column name for filtering inclusion and exclusion ie. \code{is.na(filter)}}

\item{days}{A selected time period to consider as similar injury eg. 3 days}

\item{verbose}{Keep variables for making the selection ie. xx.VariableName to produce DELXX}

\item{clean}{Delete all the possible duplicated cases}
}
\value{
A dataset with a DELXX column to indicate non-cases with value 1 if argument \code{clean = FALSE}. But if \code{verbose = TRUE}
the output will include all variables needed to create DELXX.
}
\description{
Identify cases from FMDS dataset. Three possible methods are used:
\enumerate{
\item Keep only those that are found in FMDS and those found in both FMDS and Somatic when the date and health institution are matched
\item If multiple registeries in FMDS with similar dates and health institution, then keep the earlier date
\item If mulitple registeries in FMDS with same date \strong{BUT} different health institutions, check with Somatic data. Matching
health institution will be kept.
}
}
\examples{
\dontrun{
 dd <- find_case(d1, d2)
}
}
