% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R
\name{find_cause}
\alias{find_cause}
\title{Find external causes for cases}
\usage{
find_cause(
  d1,
  d2,
  id = "lopenr",
  skade = "skadeDato",
  rhf = "helseforetak_nr",
  filter = NULL,
  days = 3,
  verbose = FALSE,
  clean = FALSE
)
}
\arguments{
\item{d1}{Dataset for FMDS ie. external explanation for injury}

\item{d2}{Dataset for NPR ie. entry registration for injury}

\item{id}{Unique patient identity}

\item{skade}{Coloumn name for date of injury}

\item{rhf}{Coloumn name for health institutions}

\item{filter}{Column name for filtering inclusion and exclusion ie. \code{is.na(filter)}}

\item{days}{A selected time period to consider as similar injury. Default is 3 days}

\item{verbose}{Keep variables for making the selection ie. xx.VariableName to produce DELXX}

\item{clean}{Delete all the possible duplicated cases}
}
\value{
A dataset with a DELXX column to indicate non-cases with value 1 if argument \code{clean = FALSE}. But if \code{verbose = TRUE}
the output will include all variables needed to create DELXX.
}
\description{
Identify unique external causes of injury cases based on FMDS dataset. Three possible methods are used:
\enumerate{
\item Keep only those that are found in FMDS, and those found in both FMDS and Somatic when the date and health institution are matched
\item If multiple registries in FMDS with similar dates and health institution identified, then keep the earlier based on time
\item If mulitple registries in FMDS with same date \strong{BUT} different health institutions identified, then check
health institution in Somatic data. Matching health institution will be kept.
}
}
\examples{
\dontrun{
 dd <- find_cause(d1, d2)
}
}
