% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyr_fy.R
\name{fy}
\alias{fy}
\title{Fiscal Year}
\usage{
fy(fy_date, fy_start = 7)
}
\arguments{
\item{fy_date}{Date in date format.}

\item{fy_start}{Must be an integer in range of 1 - 12 to represent the starting month of the fiscal year. The default is 7.}
}
\description{
Returns an integer year from a date based on the starting month of the fiscal year
}
\examples{
fy(as.Date("2019-05-01"),6)
}
\author{
Michael Valcic
}
