% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyrtaarn-package.R
\docType{package}
\name{fyrtaarn-package}
\alias{fyrtaarn}
\alias{fyrtaarn-package}
\title{fyrtaarn: Anaysis of data on injuries}
\description{
This package help in running analysis on injuries from NPR and KPR data.
}
\details{
Analysis of data on injuries from NPR and KPR data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fyrtaarn.github.io/fyrtaarn/}
}

}
\author{
\strong{Maintainer}: Yusman Kamaleri \email{ybka@fhi.no} (\href{https://orcid.org/0000-0001-5014-3665}{ORCID})

}
\keyword{internal}
