% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-codes.R
\name{get_valid_codes}
\alias{get_valid_codes}
\title{Get valid codes}
\usage{
get_valid_codes(
  d,
  select.col = "hoveddiagnoser",
  create.col = "hovdiag",
  split = " ",
  keep = FALSE
)
}
\arguments{
\item{d}{Dataset}

\item{select.col}{Columnames from dataset with ICD codes to be controlled for. Default is \code{hoveddiagnoser}}

\item{create.col}{New columnames consisting logical value to indicate that at least one
of these codes S00 to T78 exists ie. TRUE means one or more of the codes in \code{select.col}
is between S00 to T78. Default is \code{hovdiag}}

\item{split}{Symbols for splitting the codes when there are multiple codes in the column. Default is whitespace ie. \code{" "}}

\item{keep}{Split multiple codes and keep in its own column. Default is FALSE.}
}
\value{
Original column will have suffix \verb{*.old}.
}
\description{
Select codes from column consisting ICD-10 codes that meet the requirement for injury codes ie. between S00 to T78
}
\examples{
\dontrun{
d1 <- get_valid_codes(dt = dd, "hoveddiagnoser", "hovdiag")
}
}
