% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R
\name{is_dup_rhf}
\alias{is_dup_rhf}
\title{Duplicated cases from similar health institution}
\usage{
is_dup_rhf(d, id, skade, rhf, suffix = 1)
}
\arguments{
\item{d}{Dataset}

\item{id}{Unique patient identity}

\item{skade}{Coloumn name for date of injury}

\item{rhf}{Coloumn name for health institutions}

\item{suffix}{Surfix to be added to colum names for filtering eg. xx.var1}
}
\description{
Identify cases with similar date of injury and if found,
check if they are registered from the same health institutions.
Keep the first entry only with similar health institutions.
}
\keyword{internal}
