% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{is_rhf}
\alias{is_rhf}
\title{Duplicated with different health institutions}
\usage{
is_rhf(d1, d2, id, skade, rhf, filter = NULL, days = 3)
}
\arguments{
\item{d1}{Dataset for FMDS ie. external explanation for injury}

\item{d2}{Dataset for NPR ie. intery registration for injury}

\item{id}{Unique patient identity}

\item{skade}{Coloumn name for date of injury}

\item{rhf}{Coloumn name for health institutions}

\item{filter}{Column name for filtering inclusion and exclusion ie. \code{is.na(filter)}}

\item{days}{A selected time period to consider as similar injury eg. 3 days}
}
\description{
Identify cases with similar date of injury but has different health institutions.
Need to check with injury register to identify which health institutions
the patients were registered to in NPR. If the health institution in FMDS isn't
similar to those registered in NPR (entry point) within a selected time period
then it's considered double registery. Double registery is marked with 1 in column "xx.DEL"
}
\keyword{internal}
