% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-date.R
\name{is_same_dates}
\alias{is_same_dates}
\title{Same date}
\usage{
is_same_dates(d1, d2)
}
\arguments{
\item{d1}{FMDS dataset}

\item{d2}{Somatic dataset}
}
\description{
Keeping cases with similar date and health institution in FMDS and Somatic data only.
Nevertheless, patients registered in FMDS that aren't found in Somatic data will be kept.
}
