% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same-date.R
\name{is_same_dates}
\alias{is_same_dates}
\title{Same date}
\usage{
is_same_dates(
  d1,
  d2,
  key1 = c("lopenr", "skadeDato"),
  key2 = c("lopenr", "innDato")
)
}
\arguments{
\item{d1}{FMDS dataset}

\item{d2}{Somatic dataset}

\item{key1}{Columnames in \code{d1} for sorting the data. Default are \code{lopenr} and \code{skadeDato}}

\item{key2}{Columnames in \code{d2} for sorting the data. Default are \code{lopenr} and \code{innDato}}
}
\description{
Keeping cases with similar date and health institution in FMDS and Somatic data only.
Nevertheless, patients registered in FMDS that aren't found in Somatic data will be kept.
}
